MySQL（ローカルDB）前提で、/expense機能を実装（仮完成版）。 ユーザーIDの特定はコメントアウトでエスケープ。 DBデータを用いてのテスト未実施。
