+++ backend/requirements.txt
@@ -13,3 +13,6 @@ cryptography==41.0.7  # Required for pymysql
 # Migrations
 alembic==1.12.1
 
+# user_idをtokenから取得するために必要
+google-auth
+requests
